# laravel.pu012.com/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Api Project",
        "description": "Demo my Project ",
        "contact": {
            "email": "admin@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "https://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "paths": {
        "/api/products": {
            "get": {
                "tags": [
                    "Product"
                ],
                "operationId": "ab015cc2055ac75e6d23c628675c26e7",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "name",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "page_size",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List Products."
                    }
                }
            },
            "put": {
                "tags": [
                    "Product"
                ],
                "operationId": "fc0eee8e1623ba7c45c1cf287c490ce5",
                "requestBody": {
                    "description": "Create product info",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "id",
                                    "name",
                                    "detail"
                                ],
                                "properties": {
                                    "id": {
                                        "type": "integer"
                                    },
                                    "name": {
                                        "type": "string"
                                    },
                                    "detail": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "not found"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            },
            "post": {
                "tags": [
                    "Product"
                ],
                "operationId": "7ff81b1681ff3180e5b73a2e15f2c67b",
                "requestBody": {
                    "description": "Create product info",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "detail"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "detail": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "not found"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/api/products/{id}": {
            "get": {
                "tags": [
                    "Product"
                ],
                "operationId": "47affec446a4e21ba3cbf0b6b14abb44",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Buscar por estado",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List Products."
                    }
                }
            },
            "delete": {
                "tags": [
                    "Product"
                ],
                "operationId": "33173a96c7ab5ed4ba9ceaffb4c6003d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Buscar por estado",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "not found"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        }
    }
}